+++ contracts/test/strategies/base/TestBorrowStrategy.sol
@@ -337,9 +337,9 @@ contract TestBorrowStrategy is BorrowStrategy {
         deltas = new int256[](2);
     }
 
-    function testCalcOriginationFee(uint16 _origFee, uint40 emaUtilRate, uint8 minUtilRate, uint16 feeDivisor, uint256 liquidityBorrowed, uint256 borrowedInvariant, uint256 lpInvariant, uint256 discount) external virtual {
+    function testCalcOriginationFee(uint16 _origFee, uint32 emaUtilRate, uint8 minUtilRate, uint16 feeDivisor, uint256 liquidityBorrowed, uint256 borrowedInvariant, uint256 lpInvariant, uint256 discount) external virtual {
         s.origFee = _origFee; // 2 basis points
-        s.emaUtilRate = emaUtilRate; // 10 decimals
+        s.emaUtilRate = emaUtilRate; // 8 decimals
         s.minUtilRate = minUtilRate; // 85 integer
         s.feeDivisor = feeDivisor; // 16384;
